requirements.txt
imbalanced-learn==0.14.0
joblib==1.5.0
mlflow==2.22.0
numpy==2.2.5
pandas==2.2.3
plotly==6.0.1
PyYAML==6.0.2
pytest==8.3.5
scikit-learn==1.6.1
shap==0.47.2
streamlit==1.44.1
xgboost==3.0.0

